requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
entrypoints==0.3
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
pickleshare==0.7.5
Pillow==8.1.1
prometheus-client==0.9.0
prompt-toolkit==1.0.18
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
textblob==0.15.3
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wordcloud==1.8.1
zipp==3.4.0
